Fix: Version 1.0 - 클린 아키텍처 적용

+++ ArchitectureTestApp/Project/Domain/Implements/Entity/Player/ValueObject/PhysicalInfo.swift
@@ -10,5 +10,5 @@ import Foundation
 struct PhysicalInfo {
     let age: Int
     let height: Double
-    let weight: Int
+    let weight: Double
 }
